Release of version 1.3.0
Now man automatically generates the setup.py, MANIFEST.in and requirements.txt when uploading and therefore is not needed anymore the deployment phase on travis nor in the installation file. It is only a proxy between you and setuptools.
Added support for console scripts.
A few but fixes on creation and release phases